+++ pkg/validator/fullaudit.go
@@ -16,19 +16,15 @@ import (
 )
 
 // RunAudit runs a full Polaris audit and returns an AuditData object
-func RunAudit(config conf.Configuration, kubeResources *kube.ResourceProvider, onlyShowFailedTests bool) (AuditData, error) {
+func RunAudit(config conf.Configuration, kubeResources *kube.ResourceProvider) (AuditData, error) {
 	displayName := config.DisplayName
 	if displayName == "" {
 		displayName = kubeResources.SourceName
 	}
 
-	results := []Result{}
-	for _, resources := range kubeResources.Resources {
-		kindResults, err := ApplyAllSchemaChecksToAllResources(&config, resources)
-		if err != nil {
-			return AuditData{}, err
-		}
-		results = append(results, kindResults...)
+	results, err := ApplyAllSchemaChecksToResourceProvider(&config, kubeResources)
+	if err != nil {
+		return AuditData{}, err
 	}
 
 	auditData := AuditData{
@@ -46,9 +42,6 @@ func RunAudit(config conf.Configuration, kubeResources *kube.ResourceProvider, o
 		Results: results,
 	}
 	auditData.Score = auditData.GetSummary().GetScore()
-	if onlyShowFailedTests {
-		auditData.RemoveSuccessfulResults()
-	}
 	return auditData, nil
 }